+++ src/hooks/glibc/GlibcHook.cpp
@@ -157,7 +157,7 @@ bool GlibcHook::containerGlibcHasToBeReplaced() const {
 }
 
 /*
- * Use the output of "ldd --version" to obtain information about the glibc version from the host.
+ * Use the output of "ldd --version" to obtain information about the glibc version.
  * Obtaining the glibc version through the glibc.so filename is not always viable since some Linux distributions
  * (e.g. Ubuntu 21.10, Fedora 35) package the library without the version in the filename.
  * Likewise, obtaining the version from executing the glibc shared object is not reliable because some distributions
@@ -166,53 +166,42 @@ bool GlibcHook::containerGlibcHasToBeReplaced() const {
  * that would require glibc headers to be available in the container, which cannot be guaranteed, e.g. in the case
  * of a slim image.
  */
-std::tuple<unsigned int, unsigned int> GlibcHook::detectHostLibcVersion() const {
-    auto lddOutput = std::string();
-    try {
-        lddOutput = sarus::common::executeCommand(lddPath.string() + " --version");
+static std::tuple<unsigned int, unsigned int> detectLibcVersion(
+    const boost::filesystem::path& lddPath,
+    const boost::optional<std::function<void()>> preExecActions,
+    const std::string& context) {
+    auto lddOutput = std::stringstream();
+    auto lddCommand = sarus::common::CLIArguments{lddPath.string(), "--version"};
+    auto status = sarus::common::forkExecWait(lddCommand, preExecActions, {}, &lddOutput);
+    if(status != 0) {
+        auto message = boost::format("Failed to detect %s glibc version. Command %s exited with status %d")
+            % context % lddCommand % status;
+        SARUS_THROW_ERROR(message.str());
     }
-    catch (sarus::common::Error& e) {
-        SARUS_RETHROW_ERROR(e, "Failed to detect host glibc version.");
-    }
+    return hooks::common::utility::parseLibcVersionFromLddOutput(lddOutput.str());
+}
 
-    return hooks::common::utility::parseLibcVersionFromLddOutput(lddOutput);
+std::tuple<unsigned int, unsigned int> GlibcHook::detectHostLibcVersion() const {
+    return detectLibcVersion(lddPath, {}, "host");
 }
 
 /*
- * Use the output of "ldd --version" to obtain information about the glibc version from the container.
- * The same considerations made for detectHostLibcVersion() apply here.
- * Because the Glibc hook runs with root privileges, this function uses the forkExecWait() utility function to
- * drop all privileges and switch to the user identity before executing a binary from the container.
- * A file is used to store information about the ldd output, since forkExecWait() does not capture stdout.
+ * Obtain information about the glibc version from the container.
+ * Because the Glibc hook runs with root privileges, this function uses the forkExecWait() utility function
+ * to change its root directory, drop all privileges, and switch to the user identity before executing
+ * the ldd binary from the container.
  */
 std::tuple<unsigned int, unsigned int> GlibcHook::detectContainerLibcVersion() const {
-    auto glibcOutput = std::string();
-    auto lddOutputPath = sarus::common::makeUniquePathWithRandomSuffix(boost::filesystem::path{"/tmp/glibc-hook-ldd-out"});
-
-    std::function<void()> preExecActions = [this, &lddOutputPath]() {
+    std::function<void()> preExecActions = [this]() {
         if(chroot(rootfsDir.c_str()) != 0) {
             auto message = boost::format("Failed to chroot to %s: %s")
                 % rootfsDir % strerror(errno);
             SARUS_THROW_ERROR(message.str());
         }
-
         hooks::common::utility::switchToUnprivilegedProcess(userIdentity.uid, userIdentity.gid);
-        sarus::common::createFileIfNecessary(lddOutputPath);
-        sarus::common::redirectStdoutToFile(lddOutputPath);
+        sarus::common::changeDirectory("/");
     };
-
-    auto lddCommand = sarus::common::CLIArguments{"/usr/bin/ldd", "--version"};
-    auto status = sarus::common::forkExecWait(lddCommand, preExecActions);
-    if(status != 0) {
-        auto message = boost::format("Failed to detect container glibc version. /usr/bin/ldd exited with status %d")
-            % status;
-        SARUS_THROW_ERROR(message.str());
-    }
-
-    auto lddOutput = sarus::common::readFile(rootfsDir / lddOutputPath);
-    boost::filesystem::remove(rootfsDir / lddOutputPath);
-
-    return hooks::common::utility::parseLibcVersionFromLddOutput(lddOutput);
+    return detectLibcVersion("/usr/bin/ldd", preExecActions, "container");
 }
 
 void GlibcHook::verifyThatHostAndContainerGlibcAreABICompatible(